pyproject.toml
[tool.poetry]
name = "workout-tracker"
version = "0.1.0"
description = ""
authors = ["Suman <62769221+sumankrome@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
bcrypt = "^4.3.0"
mysql-connector-python = "^9.3.0"
python-jwt = "^4.1.0"
flask = "^3.1.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
